move the actionbasedchangelistener to ListeningChangeFactory
