+++ src/main/java/spoon/experimental/modelobs/FineModelChangeListener.java
@@ -16,16 +16,10 @@
  */
 package spoon.experimental.modelobs;
 
-import spoon.experimental.modelobs.action.Action;
-import spoon.experimental.modelobs.action.AddAction;
-import spoon.experimental.modelobs.action.DeleteAction;
-import spoon.experimental.modelobs.action.DeleteAllAction;
-import spoon.experimental.modelobs.action.UpdateAction;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.declaration.ModifierKind;
 import spoon.reflect.path.CtRole;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -35,31 +29,6 @@ import java.util.Set;
  * is the listener that creates the action on the model. This default listener does nothing.
  */
 public class FineModelChangeListener {
-
-	private final List<ActionBasedChangeListener> listeners = new ArrayList<>(2);
-
-	public void addModelChangeListener(final ActionBasedChangeListener listener) {
-		listeners.add(listener);
-	}
-
-	public void removeModelChangeListener(final ActionBasedChangeListener listener) {
-		listeners.remove(listener);
-	}
-
-	protected void propagateModelChange(final Action action) {
-		for (ActionBasedChangeListener listener : listeners) {
-			listener.onAction(action);
-			if (action instanceof DeleteAllAction) {
-				listener.onDeleteAll((DeleteAllAction) action);
-			} else if (action instanceof DeleteAction) {
-				listener.onDelete((DeleteAction) action);
-			} else if (action instanceof AddAction) {
-				listener.onAdd((AddAction) action);
-			} else if (action instanceof UpdateAction) {
-				listener.onUpdate((UpdateAction) action);
-			}
-		}
-	}
 
 	public void onObjectUpdate(CtElement currentElement, CtRole role, CtElement newValue, CtElement oldValue) {
 	}
